+++ packages/contracts/contracts/optimistic-ethereum/iOVM/precompiles/iOVM_L2ToL1MessagePasser.sol
@@ -12,15 +12,15 @@ interface iOVM_L2ToL1MessagePasser {
      **********/
 
     event L2ToL1Message(
-       uint _nonce,
-       address _sender,
-       bytes _data
+        uint256 _nonce,
+        address _sender,
+        bytes _data
     );
 
 
     /********************
      * Public Functions *
      ********************/
-    
+
     function passMessageToL1(bytes calldata _message) external;
 }
