+++ packages/contracts/contracts/optimistic-ethereum/libraries/rlp/Lib_RLPWriter.sol
@@ -94,11 +94,11 @@ library Lib_RLPWriter {
 
     /**
      * RLP encodes a uint.
-     * @param _in The uint to encode.
-     * @return _out The RLP encoded uint in bytes.
+     * @param _in The uint256 to encode.
+     * @return _out The RLP encoded uint256 in bytes.
      */
     function writeUint(
-        uint _in
+        uint256 _in
     )
         internal
         pure
@@ -140,8 +140,8 @@ library Lib_RLPWriter {
      * @return _encoded RLP encoded bytes.
      */
     function _writeLength(
-        uint _len,
-        uint _offset
+        uint256 _len,
+        uint256 _offset
     )
         private
         pure
@@ -155,8 +155,8 @@ library Lib_RLPWriter {
             encoded = new bytes(1);
             encoded[0] = byte(uint8(_len) + uint8(_offset));
         } else {
-            uint lenLen;
-            uint i = 1;
+            uint256 lenLen;
+            uint256 i = 1;
             while (_len / i != 0) {
                 lenLen++;
                 i *= 256;
@@ -212,16 +212,16 @@ library Lib_RLPWriter {
      * @param _len Length of memory to copy.
      */
     function _memcpy(
-        uint _dest,
-        uint _src,
-        uint _len
+        uint256 _dest,
+        uint256 _src,
+        uint256 _len
     )
         private
         pure
     {
-        uint dest = _dest;
-        uint src = _src;
-        uint len = _len;
+        uint256 dest = _dest;
+        uint256 src = _src;
+        uint256 len = _len;
 
         for(; len >= 32; len -= 32) {
             assembly {
@@ -231,7 +231,7 @@ library Lib_RLPWriter {
             src += 32;
         }
 
-        uint mask = 256 ** (32 - len) - 1;
+        uint256 mask = 256 ** (32 - len) - 1;
         assembly {
             let srcpart := and(mload(src), not(mask))
             let destpart := and(mload(dest), mask)
@@ -258,20 +258,20 @@ library Lib_RLPWriter {
             return new bytes(0);
         }
 
-        uint len;
-        uint i = 0;
+        uint256 len;
+        uint256 i = 0;
         for (; i < _list.length; i++) {
             len += _list[i].length;
         }
 
         bytes memory flattened = new bytes(len);
-        uint flattenedPtr;
+        uint256 flattenedPtr;
         assembly { flattenedPtr := add(flattened, 0x20) }
 
         for(i = 0; i < _list.length; i++) {
             bytes memory item = _list[i];
 
-            uint listPtr;
+            uint256 listPtr;
             assembly { listPtr := add(item, 0x20)}
 
             _memcpy(flattenedPtr, listPtr, item.length);
